Slave Select optional + Remove clock requirement on HW-SPI
In cases where no slave select functionality is wanted, an invalid Pin number higher than 50 may be chosen. In these case,s no slave select pin will be occupied for no functionality. 

For the hardware SPI the clock PIN was required for the reset function even for the hardware SPI. This update implements a reset function which does use a empty hardware spi push as replacement for the hardware spi. The software spi is handled seperately unchanged.